requirements.txt
dify_plugin~=0.0.1b76
google-cloud-aiplatform~=1.42.1 